+++ packages/test/test-app-express-cjs/src/shop/cart/router.module.ts
@@ -1,0 +1,7 @@
+import { EndpointsRouterModule } from 'nestjs-endpoints';
+import { CartService } from './cart.service';
+
+export default EndpointsRouterModule.register({
+  rootDirectory: 'endpoints',
+  providers: [CartService],
+});